+++ Junwon/test-code-with-architecture-main/src/test/java/com/example/demo/post/controller/PostControllerTest.java
@@ -1,4 +1,4 @@
-package com.example.demo.controller;
+package com.example.demo.post.controller;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
