+++ Junwon/test-code-with-architecture-main/src/test/java/com/example/demo/post/controller/PostCreateControllerTest.java
@@ -1,4 +1,4 @@
-package com.example.demo.controller;
+package com.example.demo.post.controller;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
