+++ Junwon/test-code-with-architecture-main/src/test/java/com/example/demo/post/service/PostServiceTest.java
@@ -1,4 +1,4 @@
-package com.example.demo.service;
+package com.example.demo.post.service;
 
 import static org.assertj.core.api.AssertionsForClassTypes.*;
