+++ Junwon/test-code-with-architecture-main/src/test/java/com/example/demo/user/controller/UserControllerTest.java
@@ -1,4 +1,4 @@
-package com.example.demo.controller;
+package com.example.demo.user.controller;
 
 import static org.assertj.core.api.AssertionsForClassTypes.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
