+++ Junwon/test-code-with-architecture-main/src/test/java/com/example/demo/user/service/UserServiceTest.java
@@ -1,4 +1,4 @@
-package com.example.demo.service;
+package com.example.demo.user.service;
 
 import static org.assertj.core.api.AssertionsForClassTypes.*;
 import static org.mockito.ArgumentMatchers.any;
